Allow customization of IdentityRole/User
UserStore is now generic over TUser and TRole to allow application
specific classes of IdentityUser and IdentityRole.

Replaced the XML mapping file for NHibernate with mapping by code
classes. This allows users to extend from the mapping classes to
provide their own custom mapping for the respective IdentityUser
and IdentityRole classes.

Provided an extension method to NHibernate's configuration class
to easily add the defined mappings to the user's configuration.